requirements.txt
numpy
pandas
scikit-learn
matplotlib
pyyaml
datasets>=2.14.0
transformers>=4.40.0
torch>=2.2.0
